pyproject.toml
[tool.poetry]
name = "pytools"
version = "0.1.0"
description = "Python工具集Web应用"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
jinja2 = "^3.1.3"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pillow = "^10.2.0"
sqlalchemy = "^2.0.25"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.6"
greenlet = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
ruff = "^0.1.14"
black = "^24.1.1"
coverage = "^7.4.0"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
